requirements.txt
pytest
pillow
wand
PyPDF2
zbar-py
numpy